% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{EA_exn_env}
\alias{EA_exn_env}
\title{Exons (coding DNA) of all the transcripts in this package's transcript set}
\description{
Exons (coding DNA) of all the transcripts in this package's transcript set
}
\details{
Environment, with transcript names as keys and exons (as \code{GenomicRanges::GRanges}) as values.
}
\author{
Dennis Kostka
}
\keyword{data}
