% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{EA_spl_env}
\alias{EA_spl_env}
\title{Splice regions of all the transcripts in this package's transcript set}
\description{
Splice regions of all the transcripts in this package's transcript set
}
\details{
Envirnonment, with transcript names as keys, and splice regions (as \code{GenomicRanges::GRangesList}) as values.
Splice regions extend 3bp into the exon and 8bp into the intron.
}
\author{
Dennis Kostka
}
\keyword{data}
