% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{EA_spl_grl}
\alias{EA_spl_grl}
\title{Splice regions of all the transcripts in this package's transcript set}
\description{
Splice regions of all the transcripts in this package's transcript set
}
\details{
Splice regions of all transcripts, as \code{GenomicRanges::GRangesList}.
}
\author{
Dennis Kostka
}
\keyword{data}
