% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{EA_txs_grl}
\alias{EA_txs_grl}
\title{Transcripts in package's transcript set}
\description{
Transcripts in package's transcript set
}
\details{
All transcripts in this package's transcript list as \code{GenomicRanges::GRangesList}.
}
\author{
Dennis Kostka
}
\keyword{data}
